requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
blurhash==1.1.4
boost==0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
Django==4.2.17
djangorestframework==3.15.2
dlib @ file:///D:/bld/dlib_1697537219106/work
exceptiongroup==1.3.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.16.1
fsspec==2025.2.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.9.1
jsonpatch==1.33
jsonpointer==3.0.0
MarkupSafe==2.1.5
Mastodon.py==2.0.0
mpmath==1.3.0
networkx==3.1
numpy @ file:///D:/bld/numpy_1687808486521/work
openai==1.93.2
opencv-python==4.11.0.86
pandas==2.0.3
pillow==10.4.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-magic-bin==0.4.14
pytz==2025.1
requests==2.32.3
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
visdom==0.2.4
websocket-client==1.8.0

# 新增：图谱可视化与中文NLP
pyvis==0.3.2
ltp==4.2.0
